Add files via upload
Welcome to Soham's Work from Home Internships.

This is an initial qualifying assignment module for Python Internship. This assignment has to be submitted in next 15 days from the date of allotment.

Initially you can upload the database queries file and get it checked from us to avoid further programming complexities. Keep uploading your completed assignment parts so that we can track the progress.

Finally upload complete project on the date of submission as per the given guidelines.

Your performance in this module will decide your eligibility for final assignment.